large refactor
introduce closed pin enum instead of the previous open one. It uses
p{pin_num} as the syntax. we parse this out so that we can convert
from actual pin numbers to the enum value. The enum also includes
an offset to be added to the fuse calculation. This pin enum
is part of the chipspec.

PTerm is now generic and takes a ChipSpec to generate. It uses
the pin enum to create an InputPin struct inside of it that is
used for tracking inversions. The number of pins is inferred from
the closed pin enum.