factor netmap into more generic netmapmany
I originally was going to make a netportmap which would map nets to
ports, but then I realized this would mostly be a copy of the first map
structure. I refactored the code into a more generic one. Now we have
non-namespaced functions that create these specialized types for maximum
flexibility. These are usually write-once so we don't have to augment
things too much.